+++ Assets/Game Core/Scripts/Core/Battle/Entities/Monsters/Monster Tracker/IMonsterTracker.cs
@@ -7,12 +7,12 @@ namespace GameCore.Battle.Monsters
         event Action OnDoAttackEvent;
         event Action<MonsterStats> OnHealthChangedEvent;
         event Action OnDiedEvent;
-        event Action<int> OnTakeDamageEvent;
+        event Action<float> OnTakeDamageEvent;
         void SendHealthChanged(MonsterStats monsterStats);
         void SendDied();
         void SendAttack();
-        void TakeDamage(int damage);
-        void SetDamage(int damage);
-        int GetDamage();
+        void TakeDamage(float damage);
+        void SetDamage(float damage);
+        float GetDamage();
     }
 }
