+++ Assets/Game Core/Scripts/Core/Battle/Entities/Monsters/MonsterStats.cs
@@ -4,7 +4,7 @@
     {
         // CONSTRUCTORS: --------------------------------------------------------------------------
 
-        public MonsterStats(int currentHealth, int maxHealth)
+        public MonsterStats(float currentHealth, float maxHealth)
         {
             CurrentHealth = currentHealth;
             MaxHealth = maxHealth;
@@ -12,7 +12,7 @@
 
         // PROPERTIES: ----------------------------------------------------------------------------
         
-        public int CurrentHealth { get; }
-        public int MaxHealth { get; }
+        public float CurrentHealth { get; }
+        public float MaxHealth { get; }
     }
 }
