+++ Assets/Game Core/Scripts/Core/Battle/Entities/Player/Player Tracker/IPlayerTracker.cs
@@ -7,14 +7,14 @@ namespace GameCore.Battle.Player
         event Action OnDoAttackEvent;
         event Action<PlayerStats> OnHealthChangedEvent;
         event Action OnDiedEvent;
-        event Action<int> OnTakeDamageEvent;
+        event Action<float> OnTakeDamageEvent;
         void SendHealthChanged(PlayerStats monsterStats);
         void SendDied();
         void SendAttack();
-        void TakeDamage(int damage);
-        void SetDamage(int damage);
-        void SetDefense(int defense);
-        int GetDamage();
-        int GetDefense();
+        void TakeDamage(float damage);
+        void SetDamage(float damage);
+        void SetDefense(float defense);
+        float GetDamage();
+        float GetDefense();
     }
 }
